requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0.0
aiomysql>=0.2.0
pymysql>=1.1.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
slowapi>=0.1.9
pydantic[email]>=2.5.0
croniter>=1.4.1
wakeonlan>=3.1.0
python-nmap>=0.7.1
